+++ src/Game/SuperRegionArray.php
@@ -17,4 +17,39 @@ class SuperRegionArray extends LoadedArray
 	{
 		return $this->getOffsets();
 	}
+
+	/**
+	 * @param integer $id
+	 *
+	 * @return SuperRegion|null
+	 */
+	public function get($id)
+	{
+		if( !$this->hasSuperRegion($id) )
+		{
+			return null;
+		}
+		return $this->offsetGet($id);
+	}
+
+	/**
+	 * @param SuperRegion $super_region
+	 *
+	 * @return $this
+	 */
+	public function addSuperRegion($super_region)
+	{
+		$this->offsetSet($super_region->getId(), $super_region);
+		return $this;
+	}
+
+	/**
+	 * @param integer $super_region_id
+	 *
+	 * @return boolean
+	 */
+	public function hasSuperRegion($super_region_id)
+	{
+		return $this->offsetExists($super_region_id);
+	}
 }