+++ tests/Game/RegionTest.php
@@ -4,7 +4,6 @@ namespace Prokki\Warlight2BotTemplate\Test\Game;
 
 use PHPUnit\Framework\TestCase;
 
-use Prokki\Warlight2BotTemplate\Game\EnvironmentFactory;
 use Prokki\Warlight2BotTemplate\Game\Region;
 use Prokki\Warlight2BotTemplate\Game\RegionArray;
 use Prokki\Warlight2BotTemplate\Game\RegionState;
@@ -25,18 +24,6 @@ class RegionTest extends TestCase
 		self::assertEquals(RegionState::class, get_class($region->»getState()));
 		self::assertEquals(RegionArray::class, get_class($region->getNeighbors()));
 		self::assertEquals(123456, $region->getId());
-	}
-
-	/**
-	 * @covers \Prokki\Warlight2BotTemplate\Game\Region::getSuperRegion()
-	 */
-	public function testSetSuperRegion()
-	{
-		$super_region = new SuperRegion(4711, 1001001);
-		$region       = EnvironmentFactory::Get()->newRegion(123456)->»assignSuperRegion($super_region);
-
-		// region has a super region
-		self::assertEquals(spl_object_hash($super_region), spl_object_hash($region->getSuperRegion()));
 	}
 
 	/**
@@ -151,6 +138,22 @@ class RegionTest extends TestCase
 
 		$region->»getState()->setArmies(2000000);
 		self::assertEquals(2000000, $region->getArmies());
+	}
+
+	/**
+	 * @covers \Prokki\Warlight2BotTemplate\Game\Region::»assignSuperRegion()
+	 * @covers \Prokki\Warlight2BotTemplate\Game\Region::getSuperRegion()
+	 */
+	public function testAssignSuperRegion()
+	{
+		$region = new Region(1);
+
+		self::assertNull($region->getSuperRegion());
+
+		$super_region = new SuperRegion(1, 5);
+		$region->»assignSuperRegion($super_region);
+
+		self::assertEquals($super_region, $region->getSuperRegion());
 	}
 
 }