+++ tests/Game/SetupMapTest.php
@@ -10,35 +10,35 @@ use Prokki\Warlight2BotTemplate\Game\SetupMap;
 class SetupMapTest extends TestCase
 {
 	/**
-	 * @covers \Prokki\Warlight2BotTemplate\Game\SetupMap::addSuperRegion()
+	 * @covers \Prokki\Warlight2BotTemplate\Game\SetupMap::addSuperRegionSetUp()
 	 */
 	public function testAddSuperRegion()
 	{
 		$map = new SetupMap();
-		$map->addSuperRegion(200, 0);
-		$map->addSuperRegion(300, 7000);
-		$map->addSuperRegion(400, 21000);
+		$map->addSuperRegionSetUp(200, 0);
+		$map->addSuperRegionSetUp(300, 7000);
+		$map->addSuperRegionSetUp(400, 21000);
 
 		$this->assertAttributeInstanceOf(LoadedArray::class, '_superRegionIds', $map);
 		$this->assertEquals([200 => 0, 300 => 7000, 400 => 21000], $this->getObjectAttribute($map, '_superRegionIds')->getArrayCopy());
 	}
 
 	/**
-	 * @covers \Prokki\Warlight2BotTemplate\Game\SetupMap::addSuperRegion()
+	 * @covers \Prokki\Warlight2BotTemplate\Game\SetupMap::addSuperRegionSetUp()
 	 */
 	public function testAddSuperRegionTwice()
 	{
 		$map = new SetupMap();
-		$map->addSuperRegion(200, 0);
+		$map->addSuperRegionSetUp(200, 0);
 
 		// add super region a second time throws an exception
 		self::expectException(InitializationException::class);
 		self::expectExceptionCode(221);
-		$map->addSuperRegion(200, 7000);
+		$map->addSuperRegionSetUp(200, 7000);
 	}
 
 	/**
-	 * @covers \Prokki\Warlight2BotTemplate\Game\SetupMap::addRegion()
+	 * @covers \Prokki\Warlight2BotTemplate\Game\SetupMap::addRegionSetUp()
 	 */
 	public function testAddRegion()
 	{
@@ -46,31 +46,31 @@ class SetupMapTest extends TestCase
 
 		$this->assertAttributeInstanceOf(LoadedArray::class, '_regionIds', $map);
 
-		$map->addRegion(1, 200);
-		$map->addRegion(2, 200);
-		$map->addRegion(3, 200);
-		$map->addRegion(4, 300);
-		$map->addRegion(5, 300);
+		$map->addRegionSetUp(1, 200);
+		$map->addRegionSetUp(2, 200);
+		$map->addRegionSetUp(3, 200);
+		$map->addRegionSetUp(4, 300);
+		$map->addRegionSetUp(5, 300);
 
 		$this->assertEquals([200 => [1, 2, 3], 300 => [4, 5]], $this->getObjectAttribute($map, '_regionIds')->getArrayCopy());
 	}
 
 	/**
-	 * @covers \Prokki\Warlight2BotTemplate\Game\SetupMap::addRegion()
+	 * @covers \Prokki\Warlight2BotTemplate\Game\SetupMap::addRegionSetUp()
 	 */
 	public function testAddRegionFailed()
 	{
 		$map = new SetupMap();
-		$map->addRegion(4, 300);
+		$map->addRegionSetUp(4, 300);
 
 		// add region a second time throws an exception
 		self::expectException(InitializationException::class);
 		self::expectExceptionCode(211);
-		$map->addRegion(4, 300);
+		$map->addRegionSetUp(4, 300);
 	}
 
 	/**
-	 * @covers \Prokki\Warlight2BotTemplate\Game\SetupMap::addNeighbors()
+	 * @covers \Prokki\Warlight2BotTemplate\Game\SetupMap::addNeighborsSetUp()
 	 */
 	public function testAddNeighbors()
 	{
@@ -78,14 +78,14 @@ class SetupMapTest extends TestCase
 
 		$this->assertAttributeInstanceOf(LoadedArray::class, '_neighborRegionIds', $map);
 
-		$map->addNeighbors(1, [2, 4]);
-		$map->addNeighbors(3, [4, 5]);
+		$map->addNeighborsSetUp(1, [2, 4]);
+		$map->addNeighborsSetUp(3, [4, 5]);
 
 		$this->assertEquals([1 => [2, 4], 3 => [4, 5]], $this->getObjectAttribute($map, '_neighborRegionIds')->getArrayCopy());
 	}
 
 	/**
-	 * @covers \Prokki\Warlight2BotTemplate\Game\SetupMap::addNeighbors()
+	 * @covers \Prokki\Warlight2BotTemplate\Game\SetupMap::addNeighborsSetUp()
 	 */
 	public function testAddNeighborsTwice()
 	{
@@ -93,16 +93,16 @@ class SetupMapTest extends TestCase
 
 		$this->assertAttributeInstanceOf(LoadedArray::class, '_neighborRegionIds', $map);
 
-		$map->addNeighbors(1, [2, 4]);
-		$map->addNeighbors(1, [2, 4]);
-		$map->addNeighbors(3, [5]);
-		$map->addNeighbors(3, [4, 5]);
+		$map->addNeighborsSetUp(1, [2, 4]);
+		$map->addNeighborsSetUp(1, [2, 4]);
+		$map->addNeighborsSetUp(3, [5]);
+		$map->addNeighborsSetUp(3, [4, 5]);
 
 		$this->assertEquals([1 => [2, 4], 3 => [5, 4]], $this->getObjectAttribute($map, '_neighborRegionIds')->getArrayCopy());
 	}
 
 	/**
-	 * @covers \Prokki\Warlight2BotTemplate\Game\SetupMap::addNeighbors()
+	 * @covers \Prokki\Warlight2BotTemplate\Game\SetupMap::addNeighborsSetUp()
 	 */
 	public function testAddEmptyNeighbors()
 	{
@@ -110,13 +110,13 @@ class SetupMapTest extends TestCase
 
 		$this->assertAttributeInstanceOf(LoadedArray::class, '_neighborRegionIds', $map);
 
-		$map->addNeighbors(6, []);
+		$map->addNeighborsSetUp(6, []);
 
 		$this->assertEquals([6 => []], $this->getObjectAttribute($map, '_neighborRegionIds')->getArrayCopy());
 	}
 
 	/**
-	 * @covers \Prokki\Warlight2BotTemplate\Game\SetupMap::addWasteland()
+	 * @covers \Prokki\Warlight2BotTemplate\Game\SetupMap::addWastelandSetUp()
 	 */
 	public function testAddWasteland()
 	{
@@ -124,15 +124,15 @@ class SetupMapTest extends TestCase
 
 		$this->assertAttributeInstanceOf(LoadedArray::class, '_wastelandIds', $map);
 
-		$map->addWasteland(2);
-		$map->addWasteland(5);
-		$map->addWasteland(6);
+		$map->addWastelandSetUp(2);
+		$map->addWastelandSetUp(5);
+		$map->addWastelandSetUp(6);
 
 		$this->assertEquals([2 => 2, 5 => 5, 6 => 6], $this->getObjectAttribute($map, '_wastelandIds')->getArrayCopy());
 	}
 
 	/**
-	 * @covers \Prokki\Warlight2BotTemplate\Game\SetupMap::addWasteland()
+	 * @covers \Prokki\Warlight2BotTemplate\Game\SetupMap::addWastelandSetUp()
 	 */
 	public function testAddWastelandTwice()
 	{
@@ -140,8 +140,8 @@ class SetupMapTest extends TestCase
 
 		$this->assertAttributeInstanceOf(LoadedArray::class, '_wastelandIds', $map);
 
-		$map->addWasteland(2);
-		$map->addWasteland(2);
+		$map->addWastelandSetUp(2);
+		$map->addWastelandSetUp(2);
 
 		$this->assertEquals([2 => 2], $this->getObjectAttribute($map, '_wastelandIds')->getArrayCopy());
 	}