+++ tests/MapTest.php
@@ -24,41 +24,36 @@ abstract class MapTest extends TestCase
 	 */
 	protected $_map = null;
 
-	public static function setUpBeforeClass()
-	{
-		EnvironmentFactory::Init();
-	}
-
 	public function setUp()
 	{
 		$this->_map = new Map();
 
-		$this->_map->addSuperRegion(1, 20);
-		$this->_map->addSuperRegion(2, 19);
-		$this->_map->addSuperRegion(3, 18);
-		$this->_map->addSuperRegion(4, 17);
+		$this->_map->addSuperRegionSetUp(1, 20);
+		$this->_map->addSuperRegionSetUp(2, 19);
+		$this->_map->addSuperRegionSetUp(3, 18);
+		$this->_map->addSuperRegionSetUp(4, 17);
 
-		$this->_map->addRegion(1, 2);
-		$this->_map->addRegion(2, 4);
-		$this->_map->addRegion(3, 1);
-		$this->_map->addRegion(4, 2);
-		$this->_map->addRegion(5, 3);
-		$this->_map->addRegion(6, 3);
-		$this->_map->addRegion(7, 4);
-		$this->_map->addRegion(8, 3);
-		$this->_map->addRegion(9, 4);
+		$this->_map->addRegionSetUp(1, 2);
+		$this->_map->addRegionSetUp(2, 4);
+		$this->_map->addRegionSetUp(3, 1);
+		$this->_map->addRegionSetUp(4, 2);
+		$this->_map->addRegionSetUp(5, 3);
+		$this->_map->addRegionSetUp(6, 3);
+		$this->_map->addRegionSetUp(7, 4);
+		$this->_map->addRegionSetUp(8, 3);
+		$this->_map->addRegionSetUp(9, 4);
 
-		$this->_map->addNeighbors(1, [4]);
-		$this->_map->addNeighbors(2, [7]);
-		$this->_map->addNeighbors(3, [4]);
-		$this->_map->addNeighbors(4, [5, 8]);
-		$this->_map->addNeighbors(5, [6, 8]);
-		$this->_map->addNeighbors(6, [7, 8]);
-		$this->_map->addNeighbors(7, [9]);
-		$this->_map->addNeighbors(8, [9]);
+		$this->_map->addNeighborsSetUp(1, [4]);
+		$this->_map->addNeighborsSetUp(2, [7]);
+		$this->_map->addNeighborsSetUp(3, [4]);
+		$this->_map->addNeighborsSetUp(4, [5, 8]);
+		$this->_map->addNeighborsSetUp(5, [6, 8]);
+		$this->_map->addNeighborsSetUp(6, [7, 8]);
+		$this->_map->addNeighborsSetUp(7, [9]);
+		$this->_map->addNeighborsSetUp(8, [9]);
 
-		$this->_map->addWasteland(2);
-		$this->_map->addWasteland(4);
+		$this->_map->addWastelandSetUp(2);
+		$this->_map->addWastelandSetUp(4);
 
 		$this->_map->initialize();
 	}